requirements.txt
numpy==1.26.4
transformers==4.40.0
torch
tensorflow
flask
pillow